requirements.txt
python-telegram-bot==21.0.1
openai==1.58.1
python-dotenv==1.0.0
requests==2.31.0
